refactor RedisTopicConsumer into RedisSubscriber
